+++ backend/users/schemas.py
@@ -15,13 +15,3 @@ class User(UserBase):
     id: int
     class Config:
         orm_mode = True
-
-
-class Token(BaseModel):
-    access_token: str
-    token_type: str
-
-
-class TokenData(BaseModel):
-    email: str = None
-    story_id: int = None
